Update 7in5 V3 implementation
