+++ src/epd7in5_v3/graphics.rs
@@ -1,14 +1,14 @@
-use crate::epd7in5_v2::{DEFAULT_BACKGROUND_COLOR, HEIGHT, WIDTH};
-use crate::graphics::{Display, DisplayRotation};
-use embedded_graphics_core::pixelcolor::BinaryColor;
+use crate::color::TriColor;
+use crate::epd7in5_v3::{DEFAULT_BACKGROUND_COLOR, HEIGHT, WIDTH, NUM_DISPLAY_BITS};
+use crate::graphics::{TriDisplay, DisplayRotation};
 use embedded_graphics_core::prelude::*;
 
 /// Full size buffer for use with the 7in5 EPD
 ///
 /// Can also be manually constructed:
-/// `buffer: [DEFAULT_BACKGROUND_COLOR.get_byte_value(); WIDTH / 8 * HEIGHT]`
+/// `buffer: [DEFAULT_BACKGROUND_COLOR.get_byte_value(); 2 * NUM_DISPLAY_BITS as usize]`
 pub struct Display7in5 {
-    buffer: [u8; WIDTH as usize * HEIGHT as usize / 8],
+    buffer: [u8; 2 * NUM_DISPLAY_BITS as usize],
     rotation: DisplayRotation,
 }
 
@@ -16,14 +16,14 @@ impl Default for Display7in5 {
     fn default() -> Self {
         Display7in5 {
             buffer: [DEFAULT_BACKGROUND_COLOR.get_byte_value();
-                WIDTH as usize * HEIGHT as usize / 8],
+                2 * NUM_DISPLAY_BITS as usize],
             rotation: DisplayRotation::default(),
         }
     }
 }
 
 impl DrawTarget for Display7in5 {
-    type Color = BinaryColor;
+    type Color = TriColor;
     type Error = core::convert::Infallible;
 
     fn draw_iter<I>(&mut self, pixels: I) -> Result<(), Self::Error>
@@ -31,7 +31,7 @@ impl DrawTarget for Display7in5 {
         I: IntoIterator<Item = Pixel<Self::Color>>,
     {
         for pixel in pixels {
-            self.draw_helper(WIDTH, HEIGHT, pixel)?;
+            self.draw_helper_tri(WIDTH, HEIGHT, pixel, crate::graphics::DisplayColorRendering::Positive)?;
         }
         Ok(())
     }
@@ -43,7 +43,7 @@ impl OriginDimensions for Display7in5 {
     }
 }
 
-impl Display for Display7in5 {
+impl TriDisplay for Display7in5 {
     fn buffer(&self) -> &[u8] {
         &self.buffer
     }
@@ -59,13 +59,40 @@ impl Display for Display7in5 {
     fn rotation(&self) -> DisplayRotation {
         self.rotation
     }
+
+    fn chromatic_offset(&self) -> usize {
+        NUM_DISPLAY_BITS as usize
+    }
+
+    fn bw_buffer(&self) -> &[u8] {
+        &self.buffer[0..self.chromatic_offset()]
+    }
+
+    fn chromatic_buffer(&self) -> &[u8] {
+        &self.buffer[self.chromatic_offset()..]
+    }
+    /*fn buffer(&self) -> &[u8] {
+        &self.buffer
+    }
+
+    fn get_mut_buffer(&mut self) -> &mut [u8] {
+        &mut self.buffer
+    }
+
+    fn set_rotation(&mut self, rotation: DisplayRotation) {
+        self.rotation = rotation;
+    }
+
+    fn rotation(&self) -> DisplayRotation {
+        self.rotation
+    }*/
 }
 
 #[cfg(test)]
 mod tests {
     use super::*;
     use crate::color::{Black, Color};
-    use crate::epd7in5_v2;
+    use crate::epd7in5_v3;
     use crate::graphics::{Display, DisplayRotation};
     use embedded_graphics::{
         prelude::*,
@@ -84,7 +111,7 @@ mod tests {
     fn graphics_default() {
         let display = Display7in5::default();
         for &byte in display.buffer() {
-            assert_eq!(byte, epd7in5_v2::DEFAULT_BACKGROUND_COLOR.get_byte_value());
+            assert_eq!(byte, epd7in5_v3::DEFAULT_BACKGROUND_COLOR.get_byte_value());
         }
     }
 
@@ -101,7 +128,7 @@ mod tests {
         assert_eq!(buffer[0], Color::Black.get_byte_value());
 
         for &byte in buffer.iter().skip(1) {
-            assert_eq!(byte, epd7in5_v2::DEFAULT_BACKGROUND_COLOR.get_byte_value());
+            assert_eq!(byte, epd7in5_v3::DEFAULT_BACKGROUND_COLOR.get_byte_value());
         }
     }
 
@@ -119,7 +146,7 @@ mod tests {
         assert_eq!(buffer[0], Color::Black.get_byte_value());
 
         for &byte in buffer.iter().skip(1) {
-            assert_eq!(byte, epd7in5_v2::DEFAULT_BACKGROUND_COLOR.get_byte_value());
+            assert_eq!(byte, epd7in5_v3::DEFAULT_BACKGROUND_COLOR.get_byte_value());
         }
     }
 
@@ -137,7 +164,7 @@ mod tests {
         assert_eq!(buffer[0], Color::Black.get_byte_value());
 
         for &byte in buffer.iter().skip(1) {
-            assert_eq!(byte, epd7in5_v2::DEFAULT_BACKGROUND_COLOR.get_byte_value());
+            assert_eq!(byte, epd7in5_v3::DEFAULT_BACKGROUND_COLOR.get_byte_value());
         }
     }
 
@@ -155,7 +182,7 @@ mod tests {
         assert_eq!(buffer[0], Color::Black.get_byte_value());
 
         for &byte in buffer.iter().skip(1) {
-            assert_eq!(byte, epd7in5_v2::DEFAULT_BACKGROUND_COLOR.get_byte_value());
+            assert_eq!(byte, epd7in5_v3::DEFAULT_BACKGROUND_COLOR.get_byte_value());
         }
     }
 }
